requirements.txt
streamlit
pandas
scikit-learn==1.6.1
joblib
pyyaml


